docs(site): rename props-of to props-docs
