+++ packages/docs/components/props-docs/index.ts
@@ -1,0 +1,1 @@
+export * from './props-docs'
