+++ packages/docs/components/props-docs/props-docs.module.css
@@ -1,4 +1,4 @@
-.propsOf {
+.propsDocs {
   border: var(--sl-border-base);
   border-radius: var(--sl-border-radius-large);
   margin-top: var(--sl-space-3);
