--- packages/docs/components/props-of/index.ts
@@ -1,1 +1,0 @@
-export * from './props-of'
