+++ packages/docs/examples/bleed.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Bleed, Button } from '@vtex/shoreline'
+
+export default function Example() {
+  return (
+    <Bleed top="$space-2" bottom="$space-2">
+      <Button>Text</Button>
+    </Bleed>
+  )
+}
