+++ packages/docs/examples/checkbox.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Checkbox } from '@vtex/shoreline'
+
+export default function Example() {
+  return <Checkbox>I agree with all the terms of service</Checkbox>
+}
