+++ packages/docs/pages/_app.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import type { AppProps } from 'next/app'
 import { GoogleAnalytics } from '@next/third-parties/google'
 import '@vtex/shoreline/css-unlayered'
+import './global.css'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
