+++ packages/docs/pages/components/alert/api-reference.mdx
@@ -18,4 +18,4 @@ Alert indicators allow users to view semantic messages that are prominent and ca
 
 ### Optional
 
-<PropsOf name="alert" />
+<PropsDocs name="alert" />
