+++ packages/docs/pages/components/bleed/api-reference.mdx
@@ -2,50 +2,10 @@
 
 Allows the content to bleed into the external container
 
-```jsx
-<Bleed top="$space-2" bottom="$space-2">
-  <Button>Text</Button>
-</Bleed>
-```
-
-## Parameters
-
-### `props`
-
-- Type: `@vtex/shoreline-components.BleedProps & @types/react.React.RefAttributes<typescript.HTMLDivElement>`
+<Preview name="bleed" />
 
 ## Props
 
-### `bottom` (optional)
+### Optional
 
-Bottom bleed
-
-- Type: `string`
-
-- Default: `'$space-0'`
-
-### `end` (optional)
-
-End bleed
-
-- Type: `string`
-
-- Default: `'$space-0'`
-
-### `start` (optional)
-
-Start bleed
-
-- Type: `string`
-
-- Default: `'$space-0'`
-
-### `top` (optional)
-
-Top bleed
-
-- Type: `string`
-
-- Default: `'$space-0'`
-
-[View source](https://github.com/vtex/shoreline/blob/7e03f7690/packages/components/src/bleed/bleed.tsx#L14)
+<PropsDocs name="bleed" />
