+++ packages/docs/pages/components/button/api-reference.mdx
@@ -22,10 +22,10 @@ Buttons triggers allow users to identify and start the most important actions in
 
 ### Required
 
-<PropsOf name="button" required />
+<PropsDocs name="button" required />
 
 ### Optional
 
-<PropsOf name="button" />
+<PropsDocs name="button" />
