+++ packages/docs/pages/components/center/api-reference.mdx
@@ -8,4 +8,4 @@ Centralizes its content
 
 ### Optional
 
-<PropsOf name="center" />
+<PropsDocs name="center" />
