+++ packages/docs/pages/components/checkbox/api-reference.mdx
@@ -2,20 +2,20 @@
 
 Checkbox controls allow users to make multiple independent choices in a form where there are at most five options and each option is binary.
 
-```jsx
-<Checkbox>Label</Checkbox>
-```
+<Preview name="checkbox" />
 
-## Parameters
+## Examples
 
-### `props`
+### States
 
-- Type: `@vtex/shoreline-components.AriaCheckboxProps & @types/react.React.RefAttributes<typescript.HTMLDivElement>`
+<Preview name="checkbox-states" />
 
-## Props
+### As Field
 
-### `CheckboxProps`
+Checkbox can be combined with [Field](/guides/forms/form-fields) components to be used within forms.
 
-- Type: `@vtex/shoreline-components.AriaCheckboxProps`
+<Preview name="checkbox-as-field" />
 
-[View source](https://github.com/vtex/shoreline/blob/7e03f7690/packages/components/src/checkbox/checkbox.tsx#L15)
+## Optional Props
+
+<PropsDocs name="checkbox" />
