+++ packages/docs/pages/global.css
@@ -1,0 +1,3 @@
+h2.nx-border-b {
+  border: none;
+}
